Task Modificación clases
Se han reestructurado las clases
