+++ CBDEclipseProyect/src/main/java/repository/IndividuoRepositorio.java
@@ -1,0 +1,15 @@
+package repository;
+
+import javax.persistence.EntityManager;
+
+public class IndividuoRepositorio {
+	
+	static EntityManager em;
+	
+	public static void createIndividuo() {
+		em.getTransaction().begin();
+		em.persist(e1);
+	}
+
+
+}
